Refined Resource Management. For HW A/V decoder and HW clock resource, use uridecodebin to parse stream info. According to these info, request related resource. For gdl plane resource, use "prepare-gdl-plane" signal of ismd_vidrend_sink to request resource.

+++ src/meego-media-player.c
@@ -82,7 +82,7 @@ struct _MeegoMediaPlayerPrivate {
   guint    timeout_id;
 
   //Media player "snapshot", for suspend/restore operation.
-  gint position;
+  gint64   position;
 
 };
